Add JSDoc types to connector code
Add JSDoc types to connector code to make it easier to spot mistakes and work with the code.

Most `tryCatchWrapper` calls were changed to different style so that the clean function definition is first, and then it is overridden with a wrapped version. This was the only way I could find to make IntelliJ IDEA and VS Code to correctly use the JSDoc for the wrapped function so that the parameter types and return types are correctly available both inside the function and when it's being called somewhere else.