Sync boolean invalidDateString instead of _inputValue
Syncing the _inputValue caused a regression:
When you have once synced the _inputValue from client to server, then
detach and attach the field, Flow tries to synchronize the _inputValue
back to client, but at that point datepicker.$ is not yet defined, and
trying to set the value to internal input throws an error.

My first approach was to use @ClientCallable instead of @Synchronize to
sync the value only from client to server via method call, but the
@ClientCallable methods are called after the event listeners, ie. after
the validation, so that doesn't work.